move configurable entry scanner task to ConfigParser
